Silence -Wshorten-64-to-32 narrowing
Widen loop counters and intermediate variables from int/unsigned to
NSInteger/NSUInteger in 14 sites. Where a public method's declared
return type is unsigned (e.g. -[FSItem childCount], -[FileKindStatistic
fileCount]), keep the type and cast at the boundary so callers don't
need to change.

Also fixes one stale int-typed parameter on the DIXTableView delegate
protocol (no implementer in the codebase, so widening to NSInteger has
no callers to update).

+++ src/AppsForItem.m
@@ -148,9 +148,9 @@
             result = [NSMutableArray arrayWithArray:(NSArray*)outURLs];
             
             // filter out .exe files
-            int i, cnt = [result count];
+            NSInteger i, cnt = [result count];
             NSURL *url;
-            
+
             for (i=(cnt-1);i>=0;i--)
             {
                 url = [result objectAtIndex:i];
